Avoid decorating built-in numeric type

+++ api/in-progress/net8.0/OpenAI.Audio.net8.0.cs
@@ -197,7 +197,7 @@ namespace OpenAI.Audio {
         public string Language { get; set; }
         public string Prompt { get; set; }
         public AudioTranscriptionFormat? ResponseFormat { get; set; }
-        public float? Temperature { get; set; }
+        public double? Temperature { get; set; }
         public AudioTimestampGranularities TimestampGranularities { get; set; }
     }
     [Experimental("OPENAI001")]
@@ -240,7 +240,7 @@ namespace OpenAI.Audio {
     public class AudioTranslationOptions : IJsonModel<AudioTranslationOptions>, IPersistableModel<AudioTranslationOptions> {
         public string Prompt { get; set; }
         public AudioTranslationFormat? ResponseFormat { get; set; }
-        public float? Temperature { get; set; }
+        public double? Temperature { get; set; }
     }
     [Experimental("OPENAI001")]
     public class DiarizedAudioTranscription : IJsonModel<DiarizedAudioTranscription>, IPersistableModel<DiarizedAudioTranscription> {
